+++ Changelog.md
@@ -8,6 +8,8 @@ Features:
  * Inline Assembly: Present proper error message when not supplying enough arguments to a functional
    instruction.
  * Inline Assembly: introduce ``keccak256`` as an opcode. ``sha3`` is still a valid alias.
+ * Inline Assembly: ``for`` and ``switch`` statements.
+ * Inline Assembly: function definitions and function calls.
 
 Bugfixes:
  * Type Checker: Make UTF8-validation a bit more sloppy to include more valid sequences.